+++ aqua/diagnostics/lat_lon_profiles/plot_lat_lon_profiles.py
@@ -2,7 +2,7 @@ import matplotlib.pyplot as plt
 
 from aqua.core.graphics import plot_seasonal_lat_lon_profiles
 from aqua.core.logger import log_configure
-from aqua.core.util import to_list, strlist_to_phrase, DEFAULT_REALIZATION
+from aqua.core.util import to_list, strlist_to_phrase, unit_to_latex, DEFAULT_REALIZATION
 from aqua.core.graphics import plot_lat_lon_profiles
 from aqua.diagnostics.base import OutputSaver, TitleBuilder