+++ aqua/diagnostics/timeseries/base.py
@@ -5,8 +5,8 @@ import xarray as xr
 from aqua.core.fixer import EvaluateFormula
 from aqua.core.logger import log_configure
 from aqua.core.util import frequency_string_to_pandas, pandas_freq_to_string
-from aqua.core.util import time_to_string, strlist_to_phrase
 from aqua.diagnostics.base import Diagnostic, start_end_dates, OutputSaver, TitleBuilder
+from aqua.core.util import time_to_string, strlist_to_phrase, unit_to_latex
 
 xr.set_options(keep_attrs=True)
 
@@ -384,8 +384,7 @@ class PlotBaseMixin():
 
         description += f'of {self.long_name} '
         if self.units is not None:
-          units = self.units.replace("**", r"\*\*")
-          description += f'[{units}] '
+          description += f'[{self.units}] '
         if self.short_name is not None:
           description += f'({self.short_name}) '